requirements.txt
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.1
django-crispy-forms==1.9.2
django-formtools==2.2
django-import-export==2.3.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.7
django-simple-captcha==0.5.12
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.2.0
pip==20.2
PyMySQL==0.10.0
pytz==2020.1
PyYAML==5.3.1
setuptools==40.8.0
six==1.15.0
tablib==2.0.0
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
